Implement a render window.
This required doing a bunch more work in Bazel.  In particular,
we had to switch to a new idiom where we build all of our
libraries as shared objects, since the render stuff really
wants to be a shared object.  We do this by first making a
library with linkstatic=1.  We then have a "cc_binary" which
is actually the shared object with linkstatic and linkshared
set to 1.  After that, when we go to depend on the library
we need to depend on the .so in the srcs, and depend on
the export of the includes in the deps.

Signed-off-by: Chris Lalancette <clalancette@osrfoundation.org>