Add in linb-any and tests.
The version of std::experimental::any that is in gcc-5 is buggy and does not
work in all of the situations we need it to.  Workaround this by using a
version of ::any from "linb", which is compatible with the ratified std::any
in C++17.  Once we move to compilers that support C++17, we can just change
the namespace and everything will be otherwise compatible.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>