Add in sleeps to the LCM channel repeater tests.
With the new threading structure in ign-transport, it can
sometimes be the case that the test for whether the handler
was properly called can happen before the handler actually
gets called, leading to test failures.  While I don't generally
like adding sleeps, this code should go away soon, so I don't
want to spend a lot of time making something more correct.
This seems to workaround the problem for me.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>