Add loadable prius simple car (#237)
* Add an additional path to the DELPHYNE_INCLUDE_DIRS.

This is so that both internal and external code can successfully
reference backend/agent_plugin_base.h.  We should eventually
clean this up, but for now this lets things work.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Add the first PriusSimpleCar loadable agent to the repository.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Move the loadable car over a bit.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Switch to using ignition instead of LCM for driving.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Pass -rdynamic.

This makes it so that the symbols we need are exported to
the loadable agents at runtime.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Set visibility back to hidden for everything except for the files we need.

It turns out that setting visibility to default for everything
is a bad idea.  Put back hidden as default, then override for
just the files we need.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Add in some documentation.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Only pass -rdynamic to the linker.

We don't need it at compile time, only at link time.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>