Re-enable and fix agent plugin loader tests.
The basic problem was that agent_plugin_loader_test.cc was
never accessing any drake symbols, so the linker was just
deciding to not link it with libdrake.so.  This is OK for
the binary, except that the loadable modules need drake
symbols to load properly.  Add the DiagramBuilder symbol
to the tests so that it gets libdrake.so properly linked in,
which allows the loadable modules to succeed.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>