[infra] run_tests always run and test against both python and cpp
Results printed at the end. Also drops the cmake call so it doesn't
override what the local environment or CI configured the build with.
If the build has not been configured by cmake, the script drops out with
a warning.

Note, it still builds/installs so that the test environment is
up to date.