[agents] rail_car shifted to be non-loadable
Also added a convenience maliput library to play around with and
for common patterns, push upstream. This means FindLane has disappeared
from AutomotiveSimulator's interface.

AgentBase added RoadGeometry to the Configure() call, necessary
for checks against the road network, but in general, the principle is
this is the point where the simulator passes to the agent the actual
ground truth world entities.