[systems] initialise from constructor, and use SetDefaultState
Previously abusing SetDefaultState by completely overriding it and
nullifying any chance to set the state/parameters with what was
actually declared at the outset.

Unfortunately LeafSystem::SetDefaultState doesn't handle the
abstract state (that appears to be have done awkwardly via
allocate instead of declare as well), so there is still some
work to be done before this class can solely rely on the parent
SetDefaultState.