[infra] disable drake info level logging (#435)
This winds back the level on drake logging. I would prefer that they had left the choice of what makes it to the screen to the application on top of drake, but that's a big change to ask and this will give us some sanity by stopping sdf_parser and multibody getting overly noisy.