[infra] bindings to python/delphyne/*.cc (#436)
* [infra] python bindings to delphyne/*.cc

* python/bindings/agents.cc -> delphyne/agents.cc
* backend/python_bindings.cc -> delphyne/bindings.cc

Was also planning to drop redundant pybind11 dependencies everywhere
(did so for the c++ tests) but found a surprising dependency in simulation_runner
that I did not notice initially because calls were prefixed with a py:: namespace.

Opted not to try and unwind that dependency (in case there was a good reason for it)
but dropped the alias py=pybind11 so it's not easy to be misled into thinking it is
a python libs method.

The 'bindings' cc module is a bit general. Could be split into simulator/maliput
bindings in the future.

* [python] maliput bindings
* [python] bindings->maliput and simulation

Split the bindings into two groups and named the modules more
specifically. Also:

* updated the agents to handle keywords args
* used pep8 style naming for methods everywhere (underscore_with_dashes)
* renamed scripts to not conflict in the bin directory with usual dash conv.

* [docs] drop the .md python example doc generator (not adding much value right now)