[infra] drop versioning for header installs
Not complete unless we consider versioning for libraries as
well and since it's not a necessary requirement in the
foreseeable future, this keeps it simpler. Easy to restore
it later if we need to.

Did keep around versioning for the CMake config and exports
though since that doesn't get in the way and doesn't need
to be touched once in place (as opposed to having to remember
versioning with every new or modified install command).