[infra] the versioning amendment (#471)
* [infra] config.hh -> version.h

Drops the unused/gui relevant macros that were getting defined as
well.

* [infra] drop versioning for header installs

Not complete unless we consider versioning for libraries as
well and since it's not a necessary requirement in the
foreseeable future, this keeps it simpler. Easy to restore
it later if we need to.

Did keep around versioning for the CMake config and exports
though since that doesn't get in the way and doesn't need
to be touched once in place (as opposed to having to remember
versioning with every new or modified install command).

* [infra] dont reinvent CMake's versioning mechanisms

* [infra] drop verbose SearchForStuff.cmake (#473)

Wasn't doing anything extra beyind find_package()
except the undesirable include_directories...

Note also, lcm has been dropped from the list to look for. The
rest have had their find_package calls moved to the root
CMakeLists.txt