Use shared_ptrs to manage the lifetimes of agents.
The main motivation for doing this is so that we can hang on
to references to the agents from the python side of things
safely.  That is, we want to make sure that both the python side
and the C++ side hold references so it doesn't get destroyed
out from either one of them.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>