Remove SetDefaultState.
Instead, if the requested speed is negative, assume that means
that the user doesn't want us to control it anymore, and
don't do anythign in CalcOutputAcceleration.  This also gives
us less work to do in the bootstrapping phase, since we won't
apply any accleration until the user has set something.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>