Split into two systems.
We now have the SpeedController, which takes two inputs
(desired velocity and feedback velocity), and one output
(acceleration), and the ConstantVectorSettable, which
takes zero inputs and has one output (the value to set).
The ConstantVectorSettable value is set based on an API
call; it is similar to ConstantVectorSource, except that
setting the value is thread-safe.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>