Fix a latent (and real) bug during destruction.
There's more information in the comments, but essentially it
turns out that we could have gotten one more ign-transport
callback while the ign-transport Node was being destroyed.
This isn't a big deal for ign-transport itself, but it *is*
a big deal for the callback.  If the callback was relying on
other members of the class that ignition::transport::Node
was embedded in, it could access them *after* those members
had been destroyed.  Because C++ rules guarantee the order
in which members are constructed/destructed, move the
ignition::transport::Node method to the end of the members,
ensuring it is the last thing constructed and the first thing
destructed.  This fixes the bug in my testing, while getting
rid of the hack that was there before.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>