Fix a latent bug in the StartStopRecording test.
It could happen that the sim_runner_->StartLogging() call
started in minute X, but that we didn't get around to
checking the string until minute X+1 (this obviously happens
more when we are near the minute boundary).  Just remove
the time portion completely, and just check that there is
a filename at all.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>