Unique id for visuals to fix rendering of model 0 (#757)
In delphyne demos with multiple agents, the model with index 0
is displayed in the wrong location. This is caused by visuals
with no specified Id being treated as having an Id of 0.
Giving a unique Id to each visual fixes the location of the
rendered model.

This also renames generate_link_id.h to generate_unique_id.h
and adds a GenerateVisualId helper function.